# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CryptoStashServer",
    "version": "v3"
  },
  "paths": {
...
    "/Wallets": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wallet"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wallet"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wallet"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Wallets"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              }
            }
          }
        }
      }
    },
    "/Wallets/{id}": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wallet"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Wallets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Wallet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Wallets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Coin": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "ticker": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MiningPool": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "poolBalances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoolBalance"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PoolBalance": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "current": {
            "type": "number",
            "format": "double"
          },
          "miningPoolId": {
            "type": "integer",
            "format": "int32"
          },
          "miningPool": {
            "$ref": "#/components/schemas/MiningPool"
          },
          "walletId": {
            "type": "integer",
            "format": "int32"
          },
          "wallet": {
            "$ref": "#/components/schemas/Wallet"
          }
        },
        "additionalProperties": false
      },
      "Wallet": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "coinId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coin": {
            "$ref": "#/components/schemas/Coin"
          },
          "workers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Worker"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Worker": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "wallet": {
            "$ref": "#/components/schemas/Wallet"
          },
          "miningPool": {
            "$ref": "#/components/schemas/MiningPool"
          }
        },
        "additionalProperties": false
      }
    }
  }
}